Fixes compilation of the VCR GUI. It couldn't possibly work before because it relied on a header that no longer exists. It adds alot of includes for standard headers though, I believe some of them are not needed (I'll fix that right now, trial and error.)

+++ main/vcr_compress.h
@@ -14,7 +14,7 @@ extern "C" {
 
 void VCRComp_init();
 
-void VCRComp_startFile( const char *filename, int width, int height, int fps );
+void VCRComp_startFile( const char *filename, long width, long height, int fps );
 void VCRComp_finishFile();
 void VCRComp_addVideoFrame( const unsigned char *data );
 void VCRComp_addAudioData( const unsigned char *data, int len );
